+++ services/preview/web/mgmt/2015-08-01-preview/web/CHANGELOG.md
@@ -1,2 +1,16 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Signature Changes
+
+#### Funcs
+
+1. ManagedApisClient.Get
+	- Params
+		- From: context.Context, string, string, *bool
+		- To: context.Context, string, string, string, *bool
+1. ManagedApisClient.GetPreparer
+	- Params
+		- From: context.Context, string, string, *bool
+		- To: context.Context, string, string, string, *bool